implimented form validation
