+++ src/app/apply/components/Experience.tsx
@@ -1,4 +1,4 @@
-import { Box, Divider, Grid, TextField, Typography } from '@mui/material'
+import { Box, Checkbox, Divider, FormControlLabel, Grid, TextField, Typography } from '@mui/material'
 import React from 'react'
 import { useDispatch, useSelector } from "react-redux";
 import type { RootState } from "@/store/store";
@@ -16,32 +16,64 @@ export default function Experience ({ errors }: { errors: any }) {
         </Typography>
         <Divider sx={{ mb: 3 }} />
         <Box sx={{ width: '100%' }}>
+            <Grid size={{xs:12,sm:6,md:4}}>
+                    <FormControlLabel
+                        control={
+                          <Checkbox
+                            checked={formData?.experience.isfresher}
+                            onChange={(e) =>
+                              handleChange("experience", "isfresher", e.target.checked)
+                            }
+                          />
+                        }
+                        label="I am a Fresher"
+                    />
+                </Grid>
             <Grid container rowSpacing={2} columnSpacing={{ xs: 1, sm: 2, md: 3 }}>
-                <Grid size={{xs:12,sm:6,md:4}}>
-                    <TextField label="Current Company"
-                    name="currentCompany"
-                    fullWidth value={formData?.experience.currentCompany}
-                    onChange={(e)=>handleChange("experience","currentCompany",e.target.value)} />
-                </Grid>
-                <Grid size={{xs:12,sm:6,md:4}}>
-                    <TextField label="Current Position"
-                    name="currentPosition"
-                    fullWidth value={formData?.experience.currentPosition}
-                    onChange={(e)=>handleChange("experience","currentPosition",e.target.value)} />
-                </Grid>
-                <Grid size={{xs:12,sm:6,md:4}}>
-                    <TextField label="Experience (in years)"
-                    name="yearsOfExperience"
-                    fullWidth value={formData?.experience.yearsOfExperience}
-                    onChange={(e)=>handleChange("experience","yearsOfExperience",e.target.value)} />
-                </Grid>
-                <Grid size={{xs:12,sm:6,md:4}}>
-                    <TextField label="I am a Fresher"
-                    name="isfresher"
-                    fullWidth value={formData?.experience.isfresher}
-                    onChange={(e)=>handleChange("experience","isfresher",e.target.value)} />
-                </Grid>
                 
+                {!formData?.experience.isfresher && (
+                <>
+                <Grid size={{xs:12,sm:6,md:4}}>
+                  <TextField
+                    label="Current Company"
+                    name="currentCompany"
+                    fullWidth
+                    value={formData?.experience.currentCompany}
+                    onChange={(e) =>
+                      handleChange("experience", "currentCompany", e.target.value)
+                    }
+                    error={Boolean(errors?.currentCompany)}
+                    helperText={errors?.currentCompany}
+                  />
+                </Grid>
+                <Grid size={{xs:12,sm:6,md:4}}>
+                  <TextField
+                    label="Current Position"
+                    name="currentPosition"
+                    fullWidth
+                    value={formData?.experience.currentPosition}
+                    onChange={(e) =>
+                      handleChange("experience", "currentPosition", e.target.value)
+                    }
+                    error={Boolean(errors?.currentPosition)}
+                    helperText={errors?.currentPosition}
+                  />
+                </Grid>
+                <Grid size={{xs:12,sm:6,md:4}}>
+                  <TextField
+                    label="Experience (in years)"
+                    name="yearsOfExperience"
+                    fullWidth
+                    value={formData?.experience.yearsOfExperience}
+                    onChange={(e) =>
+                      handleChange("experience", "yearsOfExperience", e.target.value)
+                    }
+                    error={Boolean(errors?.yearsOfExperience)}
+                    helperText={errors?.yearsOfExperience}
+                  />
+                </Grid>
+                </>
+                )}
             </Grid>
        </Box>
